+++ internal/welcome/welcome_test.go
@@ -8,7 +8,7 @@ func Test_WelcomeGenerate(t *testing.T) {
 	newUser := "Tristan";
 
 	// generate welcome message header
-	bytesOut, e := WelcomeMessageFor(newUser);
+	bytesOut, e := welcomeMessageFor(newUser);
 
 	if e != nil {
 		t.Fail();
